Revert "refactor(api): :recycle: features and trpc package consolidated into api package."
This reverts commit a7d7e31f5b36b6cc0e2662e953877762c6cf14f2.
