+++ packages/api/src/app.router.ts
@@ -1,8 +1,7 @@
 import { Injectable } from "@nestjs/common";
 
-import { GreetingRouter } from "./greeting";
-import { PostRouter } from "./post";
-import { createRouter } from "./trpc";
+import { GreetingRouter, PostRouter } from "@acme/features";
+import { createRouter } from "@acme/trpc";
 
 @Injectable()
 export class AppRouterFactory {
