+++ packages/features/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./greeting";
+export * from "./post";
