+++ packages/features/src/utils/typeHelpers.ts
@@ -1,0 +1,1 @@
+export type ObjectValues<T> = T[keyof T];
