+++ packages/trpc/src/index.ts
@@ -1,0 +1,14 @@
+import type { inferRouterInputs, inferRouterOutputs } from "@trpc/server";
+import type { Context, Session } from "context";
+import { TRPCError } from "@trpc/server";
+import { createInnerContext } from "context";
+
+export * from "trpc";
+export {
+  TRPCError,
+  Context,
+  Session,
+  createInnerContext,
+  inferRouterInputs,
+  inferRouterOutputs,
+};
