+++ packages/trpc/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "@acme/tsconfig/nest.json",
+  "compilerOptions": {
+    "baseUrl": "./src",
+    "rootDir": "./src",
+    "outDir": "./dist"
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "build", "dist"]
+}
